refactor: kill runCommand (#1267)
