+++ src/lib/spiffy/user_config_spf.ts
@@ -1,7 +1,7 @@
 import * as t from '@withgraphite/retype';
 import { execSync } from 'child_process';
 import { getGitEditor, getGitPager } from '../git/git_editor';
-import { CommandFailedError } from '../utils/run_command';
+import { CommandFailedError } from '../git/runner';
 import { spiffy } from './spiffy';
 
 const schema = t.shape({
